# Use concrete types for stream, this allows these to be stack allocated