--- a/stream/interface.go
+++ b/stream/interface.go
@@ -1,14 +1,16 @@
 package stream
 
+import "github.com/jacksontj/dataman/record"
+
 type ClientStream interface {
-	Recv() (Result, error)
+	Recv() (record.Record, error)
 	// Close completes receiving of items
 	Close() error
 }
 
 type ServerStream interface {
 	// Methods to send items
-	SendResult(Result) error
+	SendResult(record.Record) error
 	SendError(error) error
 	// Close completes sending of items-- no more items can be sent after Close() is called
 	Close() error