--- a/stream/local/all_test.go
+++ b/stream/local/all_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"testing"
 
+	"github.com/jacksontj/dataman/record"
 	"github.com/jacksontj/dataman/stream"
 )
 
@@ -11,7 +12,7 @@ import (
 func TestLocalStreams(t *testing.T) {
 	f := func(ctx context.Context) (stream.ServerStream, stream.ClientStream) {
 
-		resultsChan := make(chan stream.Result, 1)
+		resultsChan := make(chan record.Record, 1)
 		errorChan := make(chan error, 1)
 
 		server := NewServerStream(ctx, resultsChan, errorChan)